+++ shared-libs/crates/start-core/src/net/net_controller.rs
@@ -279,6 +279,7 @@ impl NetController {
                         suffix: String::new(),
                     },
                     interface_type: ServiceInterfaceType::Ui,
+                    preferred_launcher_address: None,
                 };
                 db.as_public_mut()
                     .as_server_info_mut()