+++ shared-libs/ts-modules/start-core/lib/interfaces/ServiceInterfaceBuilder.ts
@@ -26,6 +26,7 @@ export class ServiceInterfaceBuilder {
       query: Record<string, string>
       schemeOverride: { ssl: Scheme; noSsl: Scheme } | null
       masked: boolean
+      preferredLauncherAddress?: string | null
     },
   ) {}
 }