+++ shared-libs/ts-modules/start-core/lib/osBindings/ServiceInterface.ts
@@ -10,4 +10,8 @@ export type ServiceInterface = {
   masked: boolean
   addressInfo: AddressInfo
   type: ServiceInterfaceType
+  /**
+   * The interface address Open UI should prefer.
+   */
+  preferredLauncherAddress?: string | null
 }